// src/stores/types/productTypes.ts
export interface Product {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
    rating: { rate: number; count: number };
}

export interface ProductState {
    loading: boolean;
    error: string | null;
    products: Product[];
}